requirements.txt
cffi==1.14.4
future==0.18.2
kociemba==1.2.1
numpy==1.20.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pycparser==2.20
pydocstyle==5.1.1
python-i18n==0.3.9
snowballstemmer==2.1.0
